refactor: StoneColor 클래스를 분리
