+++ src/main/kotlin/omok/model/StoneColor.kt
@@ -1,0 +1,14 @@
+package omok.model
+
+enum class StoneColor {
+    BLACK,
+    WHITE,
+    ;
+
+    fun reverse(): StoneColor {
+        return when (this) {
+            BLACK -> WHITE
+            WHITE -> BLACK
+        }
+    }
+}
